feat(connector): filter ingested events by data property on string exact matches
