+++ pkg/models/meter.go
@@ -49,44 +49,6 @@ func (MeterAggregation) IsValid(input string) bool {
 	return false
 }
 
-type MeterFilterOperator string
-
-const (
-	// MeterFilterOperatorIn        MeterFilterOperator = "IN"
-	// MeterFilterOperatorNotIn     MeterFilterOperator = "NOT IN"
-	MeterFilterOperatorEquals    MeterFilterOperator = "EQ"
-	MeterFilterOperatorNot       MeterFilterOperator = "NEQ"
-	MeterFilterLowerThan         MeterFilterOperator = "LT"
-	MeterFilterLowerThanOrEq     MeterFilterOperator = "LTE"
-	MeterFilterGreaterThan       MeterFilterOperator = "GT"
-	MeterFilterGreaterThanOrEq   MeterFilterOperator = "GTE"
-	MeterFilterOperatorIsNull    MeterFilterOperator = "IS NULL"
-	MeterFilterOperatorIsNotNull MeterFilterOperator = "IS NOT NULL"
-)
-
-type MeterFilter struct {
-	Property string              `json:"property" yaml:"property"`
-	Operator MeterFilterOperator `json:"operator" yaml:"operator"`
-	Value    string              `json:"value" yaml:"value"`
-}
-
-// Values provides list valid values for Enum
-func (MeterFilterOperator) Values() (kinds []string) {
-	for _, s := range []MeterFilterOperator{
-		MeterFilterOperatorEquals,
-		MeterFilterOperatorNot,
-		MeterFilterLowerThan,
-		MeterFilterLowerThanOrEq,
-		MeterFilterGreaterThan,
-		MeterFilterGreaterThanOrEq,
-		MeterFilterOperatorIsNull,
-		MeterFilterOperatorIsNotNull,
-	} {
-		kinds = append(kinds, string(s))
-	}
-	return
-}
-
 type WindowSize string
 
 const (
@@ -146,8 +108,7 @@ type Meter struct {
 	ValueProperty string            `json:"valueProperty,omitempty" yaml:"valueProperty,omitempty"`
 	GroupBy       map[string]string `json:"groupBy,omitempty" yaml:"groupBy,omitempty"`
 	WindowSize    WindowSize        `json:"windowSize,omitempty" yaml:"windowSize,omitempty"`
-	// TODO: add filter by
-	// FilterBy      []MeterFilter
+	FilterBy      []MeterFilter     `json:"filterBy,omitempty" yaml:"filterBy,omitempty"`
 }
 
 type MeterOptions struct {
